actions: Split out zerver.actions.uploads.
